+++ src/components/_private/Flyout/FlyoutTrigger.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import React from "react";
 import type * as T from "./Flyout.types";
 import { useFlyoutContext } from "./Flyout.context";